score_leader.sma
This plug-in will show in player's chat the player with highest score. When this player's score is surpassed the new high score and the new leader of score board will be shown in chat. When a leader disconnect or resets his score the next high score will become leader and will be shown in chat.